v0.2.2.1

+++ src/api/java/appeng/api/networking/crafting/ICraftingCPU.java
@@ -52,7 +52,7 @@ public interface ICraftingCPU extends IBaseMonitor<IAEItemStack>
 	/**
 	 * @return the number of co-processors in the CPU.
 	 */
-	int getCoProcessors();
+	long getCoProcessors();
 
 	/**
 	 * @return an empty string or the name of the cpu.
